# src/main/java/sky/pro/java/diplomproject/ProjectMarketPlace/dto/UserDto.java
package sky.pro.java.diplomproject.ProjectMarketPlace.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class UserDto {

    private Integer id;
    private String firstName;
    private String lastName;
    private String email;
    private String phone;
    private String image;

}